+++ ChangeLog.md
@@ -1,4 +1,8 @@
 # HorizonsBot Change Log
+#### HorizonsBot Version 2.6.1:
+- Split the `/list` command into subcommands
+- Added command links to `/press-kit`, `/list clubs`, and `/list petitions`
+
 #### HorizonsBot Version 2.6.0:
 - New command: `/club-skip-meeting` allows club leaders and moderators to cancel the reminder for a club's upcoming meeting and advance the next meeting timestamp if repeated meetings